Use env keys for the API's and setup at startup
This pull request is to reproduce what we had working after the hackathone. It also gets us to a place where we have better instructions on what API's and keys we need to configure to produce a functional development environment.

I was having issues with startup here: https://github.com/wenlock/RedSolidaria/issues/31. I'm hoping this resolves this issue by us getting more consistent on how auth works with the service using Google API Key's instead.

There are still some WIP items here that need to be taken care of:

- [ ] Identify the exact Google API services we need to enable to get the app functional.

Closes https://github.com/wenlock/RedSolidaria/issues/31